futher adjustments - mark logging functions as obsolete - remove tracing of MT940 files (if required one can just activate to write MT940 files (*.STA) and analyze these files) - create a log file with datetime - fix formatted trace
